CameraV3 And Object Pools
(Hopefully) the final fix for the camera
Removed Post processing effects and added in a shader for the monochrome effect.
Changed spawning of obstacles to be based on object pools instead of constantly deleting and creating new prefabs.
Added backbone for turning background and floor into pools. Not actually implemented because doing so creates a bug where occaisonally the floor/wall will move faster than intended.